+++ doc/versionHistory.rst
@@ -12,6 +12,7 @@ Version History
 
 * Update getModel in plotDonutFits to use new danish parameters saved in butler metadata to create model image instead of reoptimizing.
 * Allow getModel in plotDonutFits to be used easily in interactive mode with the butler metadata.
+* Change CWFS test to run danish in order to properly test danish outputs.
 
 .._lsst.ts.donut.viz-2.4.2